requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
openai>=1.0.0

# Groq API
groq>=0.4.0

# MongoDB
pymongo>=4.0.0
certifi

# PDF and Image Processing
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=9.1.0
pdfplumber>=0.10.0

# RAG System
chromadb>=0.4.0
rank-bm25>=0.2.2

# Scientific Computing
numpy>=1.24.0

# HTTP Requests
requests>=2.28.0

# GUI
tk

# Web API
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# Optional but recommended
tqdm
